linux_host/requirements.txt
pyserial
psutil
gpustat